pyproject.toml
[project]
name = "university-notes-assistant"
version = "0.1.0"
description = "A simple product-style University Notes Assistant using FastAPI, LangChain, Chroma, and Docker."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "jinja2>=3.1.4",
  "python-multipart>=0.0.9",
  "python-dotenv>=1.0.1",
  "langchain>=0.3.0",
  "langchain-community>=0.3.0",
  "langchain-openai>=0.2.0",
  "langchain-chroma>=0.1.0",
  "langchain-text-splitters>=0.3.0",
  "chromadb>=0.5.0",
  "pypdf>=5.0.0"
]

[tool.uv]
package = false
